Compilation error

+++ ws_wrapper/ws_local_server/ws_group.go
@@ -249,7 +249,7 @@ func (wsRouter *WsFuncRouter) GetGroupMemberListByJoinTimeFilter(input, operatio
 		return
 	}
 	userWorker.Group().GetGroupMemberListByJoinTimeFilter(&BaseSuccessFailed{runFuncName(), operationID, wsRouter.uId},
-		m["groupID"].(string), int32(m["offset"].(float64)), int32(m["count"].(float64)), int32(m["joinTimeBegin"].(float64)), int32(m["joinTimeEnd"].(float64)), operationID)
+		m["groupID"].(string), int32(m["offset"].(float64)), int32(m["count"].(float64)), int64(m["joinTimeBegin"].(float64)), int64(m["joinTimeEnd"].(float64)), operationID)
 }
 
 func (wsRouter *WsFuncRouter) GetGroupMembersInfo(input, operationID string) { //(groupId string, userList string, callback Base) {
